- fix a few positioning problems when the font size is larger than 100% - inform the user when searching via partial lists isn't (fully) supported
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@1576 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ DirectoryListing.cpp
@@ -1207,7 +1207,7 @@ bool DirectoryListing::nextResult(bool prev) {
 		}
 		advance(curResult, -1);
 	} else {
-		if (distance(searchResults.begin(), curResult) == searchResults.size()-1) {
+		if (static_cast<size_t>(distance(searchResults.begin(), curResult)) == searchResults.size()-1) {
 			return false;
 		}
 		advance(curResult, 1);
